Use ConvertUniqueKeyToPrimaryKey for cloud_migration_snapshot_partition table.
